+++ tools/tpm2_createprimary.c
@@ -60,7 +60,7 @@ struct tpm_createprimary_ctx {
     char *context_file;
     struct {
         UINT8 P :1;
-        UINT8 K :1;
+        UINT8 k :1;
     } flags;
     char *parent_auth_str;
     char *key_auth_str;
@@ -87,8 +87,8 @@ static bool on_option(char key, char *value) {
         ctx.flags.P = 1;
         ctx.parent_auth_str = value;
         break;
-    case 'K': {
-        ctx.flags.K = 1;
+    case 'k': {
+        ctx.flags.k = 1;
         ctx.key_auth_str = value;
     } break;
     case 'g': {
@@ -147,7 +147,7 @@ bool tpm2_tool_onstart(tpm2_options **opts) {
     const struct option topts[] = {
         { "hierarchy",            required_argument, NULL, 'a' },
         { "auth-hierarchy",       required_argument, NULL, 'P' },
-        { "auth-object",          required_argument, NULL, 'K' },
+        { "auth-object",          required_argument, NULL, 'k' },
         { "halg",                 required_argument, NULL, 'g' },
         { "kalg",                 required_argument, NULL, 'G' },
         { "out-context",          required_argument, NULL, 'o' },
@@ -155,7 +155,7 @@ bool tpm2_tool_onstart(tpm2_options **opts) {
         { "object-attributes",    required_argument, NULL, 'A' },
     };
 
-    *opts = tpm2_options_new("A:P:K:g:G:o:L:a:", ARRAY_LEN(topts), topts,
+    *opts = tpm2_options_new("A:P:k:g:G:o:L:a:", ARRAY_LEN(topts), topts,
             on_option, NULL, TPM2_OPTIONS_SHOW_USAGE);
 
     return *opts != NULL;
@@ -175,7 +175,7 @@ int tpm2_tool_onrun(TSS2_SYS_CONTEXT *sapi_context, tpm2_option_flags flags) {
         }
     }
 
-    if (ctx.flags.K) {
+    if (ctx.flags.k) {
         TPMS_AUTH_COMMAND tmp;
         result = tpm2_auth_util_from_optarg(sapi_context, ctx.key_auth_str, &tmp, NULL);
         if (!result) {